% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{gvector-class}
\alias{gvector}
\alias{gvector-class}
\title{Generalized vector for R}
\description{
Generalized vector for R
}
\section{Slots}{

\describe{
\item{\code{vec}}{List of elements of the vector/matrix (elements can be of different types)}

\item{\code{dim}}{Vector of dimensions of the vector/matrix (prod(dim) = length(vec))}
}}

