% Generated by roxygen2 (4.0.2): do not edit by hand
\name{new.gvector}
\alias{new.gvector}
\title{Creates a gvector out of a list}
\usage{
new.gvector(vec = list(), dim = c())
}
\arguments{
\item{vec}{List to be converted into a gvector}

\item{dim}{Vector of dimensions (like dim in a matrix). You should ensure that prod(dim) = length(vec).}
}
\description{
Creates a gvector out of a list
}

