% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gvector.R
\name{V}
\alias{V}
\title{Convinience function for creating a gvector}
\usage{
V(...)
}
\arguments{
\item{...}{All the things to put in a gvector}
}
\description{
Convinience function for creating a gvector
}

