% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gvector.R
\name{as.gvector}
\alias{as.gvector}
\title{Converts an object to gvector}
\usage{
as.gvector(vec = list(), dim = c())
}
\arguments{
\item{vec}{Object to be converted}

\item{dim}{Vector of dimensions (like dim in a matrix).}
}
\description{
Converts an object to gvector
}

