% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gvector.R
\docType{package}
\name{gvector-package}
\alias{gvector-package}
\title{Package for generalized vector in R}
\description{
Generalized vector is just a list, but with the arthmetic of a vector or a matrix
}
\details{
The only function you're likely to need from \pkg{gvector} is
\code{\link{new.gvector}}.
}

