% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{lag,gvector-method}
\alias{lag,gvector-method}
\title{Apply lag to all elements of a gvector}
\usage{
\S4method{lag}{gvector}(x, dx, drop = T, ...)
}
\arguments{
\item{x}{gvector}
}
\description{
Applied the lag functions to all elements of a gvector
}

